src/main/java/au/com/rainmore/datastructure/linkedLists/No142LinkedListCycleII.java
package au.com.rainmore.datastructure.linkedLists;

import au.com.rainmore.datastructure.extra.data.ListNode;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

/**
 * <a href="https://leetcode.com/problems/linked-list-cycle-ii/description/">
 *     142. Linked List Cycle II</a>
 *
 *
 */
public class No142LinkedListCycleII {

    private Logger logger = LoggerFactory.getLogger(No21MergeTwoSortedLists.class);

    public ListNode detectCycle(ListNode head) {
        if (head == null || head.next == null) return null;
        ListNode slow = head, fast = head;
        while (fast != null && fast.next != null) {
            slow = slow.next;
            fast = fast.next.next;

            if (slow == fast) {
                slow = head;
                while (slow != fast) {
                    slow = slow.next;
                    fast = fast.next;
                }
                return slow;
            }
        }

        return null;
    }


}
